refactor(vendo): wire route table — connections (kill-list B4)
